chenged video obj

+++ app/src/main/java/com/alesno/testtaskispring/common/PlayerStartInfo.kt
@@ -2,5 +2,5 @@ package com.alesno.testtaskispring.common
 
 class PlayerStartInfo(
     var isVideoStarted: Boolean,
-    var progressTime: Long
+    var progressTime: Int
 )
